finished register, and login. Having issues with the passwordButtonHandler and events triggering out of order
