+++ src/api/index.js
@@ -1,6 +1,6 @@
 export { getAllUsers } from './admin'
 
-export { addUser, updateUser, loginUser} from './users'
+export { addUser, updateUser, loginUser, guestUser } from './users'
 export { getProductsByQuery, addNewProduct } from './products'
 export { addNewCart } from './orders';
